move couchjs-node from src to extra
Since the build of couchjs-node is optional we should put it outside of
the src/ folder and move it to an `extra` folder (and the same for all
extras). The reasoning behind is that we should separate more clearly
what is an addon from what is the core product.

fix #COUCHDB-1989